Split smart playlist functionality into utility class
